Add level enum to TaxRates
We unfortunately need to introduce the concept of order-level tax rates
to handle the Colorado Delivery fee. (And maybe other similar things?)
The simplest solution is to add an enum to our TaxRate model to help
distinguish between rates that should be applied on an order-level and
our regular item-level taxes.

Co-Authored-By: Andrew Stewart <andrew@super.gd>